requirements.txt
openai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.1
requests>=2.32.0
feedparser>=6.0.11
jinja2>=3.1.4
python-dateutil>=2.9.0
pydantic>=2.8.2